pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["Your Name"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "0.3.1"
pydantic = "^2.10.0"
langchain-openai = "^0.2.0"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"
streamlit = "^1.32.0"
watchdog = "^4.0.0"
python-multipart = "^0.0.9"
pymupdf = "^1.23.26"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
docx2txt = "^0.8"
pymilvus = "^2.4.9"
langchain-milvus = "^0.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
